erase keystore_manager
