--- preptools/utils/keystore_manager.py
@@ -1,27 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2018 ICON Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import re
-
-
-def validate_password(password) -> bool:
-    """Verify the entered password.
-
-    :param password: The password the user entered. type(str)
-    :return: bool
-    True: When the password is valid format
-    False: When the password is invalid format
-    """
-
-    return bool(re.match(r'^(?=.*\d)(?=.*[a-zA-Z])(?=.*[?!:\.,%+-/*<>{}\(\)\[\]`"\'~_^\\|@#$&]).{8,}$', password))
